pyproject.toml
[project]
name = "macoslocalapi"
version = "0.1.0"
description = "Local MLX model served via an OpenAI-compatible FastAPI API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "mlx>=0.30.6",
    "uvicorn>=0.40.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "mlx-lm>=0.30.7",
    "piper-tts>=1.4.1",
    "mlx-audio-plus>=0.1.6",
]

[tool.pytest.ini_options]
pythonpath = ["."]
